backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ethanlchristensen <ethanlchristensen@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
requests = "^2.32.3"
ollama = "^0.4.0"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-allauth = "^64.2.1"
django-rest-auth = "^0.9.5"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-rest-knox = "^5.0.1"
python-decouple = "^3.8"
pytz = "^2024.2"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
python-weather = "^2.0.7"
anthropic = "^0.46.0"
google-genai = "^1.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
